Fix FlashAttention fp32 error: cast MambaBlock output to input dtype
Mamba CUDA selective scan can return fp32 due to float() D parameter.
This propagates to attention layers where FlashAttention rejects fp32.
Cast output back to input dtype (bf16) before residual add.

+++ train_gpt.py
@@ -805,7 +805,7 @@ class MambaBlock(nn.Module):
             y = y * F.silu(z)
 
         y = self.out_proj(y)
-        return residual + y
+        return residual + y.to(residual.dtype)
 
     @torch.compiler.disable
     def _selective_scan(self, x: Tensor, dA: Tensor, dB: Tensor, C: Tensor, D: Tensor) -> Tensor:
